refactor: cleanup sorting and aggregations
