+++ enterprise/frontend/src/embedding-sdk-package/hooks/public/use-metabase-query/bundle.ts
@@ -1,0 +1,7 @@
+import { getWindow } from "embedding-sdk-shared/lib/get-window";
+
+export const getEmbeddingSdkBundle = () =>
+  getWindow()?.METABASE_EMBEDDING_SDK_BUNDLE;
+
+export const getResolveDatasetQueryFromBundle = () =>
+  getEmbeddingSdkBundle()?.resolveDatasetQuery;
